fix: Migrate dice post processor to encapsulated class
