+++ src/view/view.ts
@@ -40,13 +40,15 @@ export default class DiceView extends ItemView {
 
         this.addChild(this.#icons);
 
-        this.plugin.app.workspace.on(
-            "dice-roller:new-result",
-            (roller: StackRoller) => {
-                if (this.plugin.data.addToView) {
-                    this.addResult(roller);
+        this.registerEvent(
+            this.plugin.app.workspace.on(
+                "dice-roller:new-result",
+                (roller: StackRoller) => {
+                    if (this.plugin.data.addToView) {
+                        this.addResult(roller);
+                    }
                 }
-            }
+            )
         );
     }
     async onOpen() {
